Split camera_web_test into smaller integration test files
