--- .ci/targets/web_platform_tests_camera_web_wasm.yaml
@@ -1,17 +1,0 @@
-tasks:
-  - name: prepare tool
-    script: .ci/scripts/prepare_tool.sh
-    infra_step: true # Note infra steps failing prevents "always" from running.
-  - name: download Dart deps
-    script: .ci/scripts/tool_runner.sh
-    args: ["fetch-deps", "--web", "--supporting-target-platforms-only"]
-    infra_step: true
-  - name: drive examples
-    script: .ci/scripts/tool_runner.sh
-    args: [
-      "drive-examples",
-      "--web",
-      "--wasm",
-      "--run-chromedriver",
-      "--packages=camera_web",
-    ]
