+++ packages/camera/camera_web/example/test_driver/integration_test.dart
@@ -4,4 +4,4 @@
 
 import 'package:integration_test/integration_test_driver.dart';
 
-Future<void> main() => integrationDriver(timeout: const Duration(minutes: 30));
+Future<void> main() => integrationDriver();
